% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankICC3levels.R
\name{rankICC3levels}
\alias{rankICC3levels}
\title{Rank-based ICC with three hierarchies}
\usage{
rankICC3levels(
  x,
  level2,
  level3,
  weights = c("level1", "level2", "level3"),
  conf.int = 0.95,
  fisher = FALSE
)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{level2}{a index vector of the level-2 units.}

\item{level3}{a index vector of the level-3 units.}

\item{weights}{a character string indicating which weighting method is used. Or an optional vector of user-defined weights to be used. Should be one of the strings \code{"level1"}, \code{"level2"}, \code{"level3"}, or a numeric vector. See Details.}

\item{conf.int}{numeric specifying confidence interval coverage.}

\item{fisher}{logical indicating whether to apply fisher transformation to compute confidence intervals.}
}
\value{
a matrix with two rows. The first row is for rank-based ICC at level 2 and the second row is for rank-based ICC at level 3. Each row with following components.
\tabular{ll}{
  \code{rankICC} \tab the rank-based ICC. \cr
  \tab \cr
  \code{SE} \tab the standard error. \cr
  \tab \cr
  \code{Lower, Upper} \tab the lower and upper bound of the confidence interval.\cr
}
}
\description{
\code{rankICC3levels} computes the rank-based intraclass correlation coefficient (ICC) with three hierarchies. Starting from the innermost level, the three levels are named level 1, level 2, and level 3. The rank-based ICC at level 2 evaluates the correlation between a random pair from the same level-2 unit. The rank-based ICC at level 3 evaluates the correlation between a random pair from the same level-3 unit but different level-2 units.
}
\details{
\code{"level1"} assigns equal weights to level-1 units; \eqn{p_{ijk}=1/(\sum_{i=1}^n\sum_{j=1}^{n_i}m_{ij})}, where \eqn{n} is the number of level-3 units, \eqn{n_i} is the number of level-2 units in the level-3 unit, and \eqn{m_{ij}} is the number of level-1 units in the level-2 unit. \code{"level2"} assigns equal weights to level-2 units; \eqn{p_{ijk}=1/(m_{ij}\sum_{i=1}^n n_{i})}. \code{"level3"} assigns equal weights to level-3 units; \eqn{p_{ijk}=1/(nn_{i}m_{ij})}.
}
\examples{
\dontrun{
rankICC3levels(x, level2, level3, weights = "level2")
}
}
